+++ src/components/core/MultiSelect.tsx
@@ -1,0 +1,5 @@
+const MultiSelect = () => {
+    return <></>
+}
+
+export default MultiSelect